requirements.txt
flask==2.2.3
flask-cors==3.0.10
python-dotenv==1.0.0
pyodbc==4.0.39
pyjwt==2.6.0
bcrypt==4.0.1
